Basic user interaction implemented
Implemented some of the basic actions to work with the user interface.
Not all actions are implemented yet. You can't delete or share lists yet. These are planned for later.
This implementation is meant to check if the user interaction is any good to us.
Backend integration will follow soon.